requirements.txt
Flask==3.0.3
imbalanced-learn==0.12.3
matplotlib==3.9.2
nltk==3.8.1
pandas==2.2.3
protobuf==5.28.3
python-dotenv==1.0.1
Sastrawi==1.0.1
scikit-learn==1.3.2
google-cloud-storage==2.10.0